pyproject.toml
[tool.poetry]
name = "onchain"
version = "0.1.0"
description = ""
authors = ["koksang <koksanggl@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.10.9"                                # matching ray-cluster deployment python version
google-cloud-bigquery = "^3.3.3"
google-cloud-bigquery-storage = "^2.16.0"
loguru = "^0.6.0"
tqdm = "^4.64.1"
hydra-core = "^1.3.0"
pathlib = "^1.0.1"
pytest = "^7.2.0"
db-dtypes = "^1.0.5"
web3 = "^5.31.0"
protobuf = "^3.20.0"
ray = { extras = ["default"], version = "^2.3.1" }
pulsar-client = "^3.1.0"
google-api-python-client = "^2.70.0"
jupyterlab = "^3.6.1"
pydantic = "^1.10.7"

[tool.ruff]
line-length = 88

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
